Renames android_gingerbread_mr1 to android_10.
