+++ android_10/build.gradle
@@ -26,7 +26,7 @@ android {
 
   defaultConfig {
     minSdkVersion 10
-    packageName "org.ros.android.android_gingerbread_mr1"
+    packageName "org.ros.android.android_10"
     targetSdkVersion 10
     versionCode 1
     versionName "1.0"
