+++ android_10/src/main/java/org/ros/android/BitmapFromCompressedImage.java
