+++ android_10/src/main/java/org/ros/android/BitmapFromImage.java
