+++ android_10/src/main/java/org/ros/android/view/camera/CompressedImagePublisher.java
