+++ android_10/src/main/res/drawable-hdpi/icon.png
