+++ android_10/src/main/res/drawable-ldpi/icon.png
