+++ android_10/src/main/res/drawable-mdpi/icon.png
