+++ android_10/src/main/res/values/common_strings.xml
