+++ android_10/src/main/res/values/styles.xml
