+++ android_acm_serial/build.gradle
@@ -16,7 +16,7 @@
 
 dependencies {
   compile project(':android_honeycomb_mr2')
-  compile project(':android_gingerbread_mr1')
+  compile project(':android_10')
 }
 
 apply plugin: 'android-library'
