+++ android_honeycomb_mr2/build.gradle
@@ -17,7 +17,7 @@
 dependencies {
   compile 'com.android.support:support-v4:18.0.+'
   compile 'org.ros.rosjava_core:rosjava_geometry:0.1.+'
-  compile project(':android_gingerbread_mr1')
+  compile project(':android_10')
 }
 
 apply plugin: 'android-library'
