+++ android_tutorial_camera/build.gradle
@@ -15,7 +15,7 @@
  */
 
 dependencies {
-  compile project(':android_gingerbread_mr1')
+  compile project(':android_10')
 }
 
 apply plugin: 'android'
