+++ docs/src/main/sphinx/overview.rst
@@ -12,7 +12,7 @@ ROS-enabling Android applications
 
 android_core provides `Android Library Projects`_ to help you write ROS
 applications for Android. The library projects are named for the `Android API
-level`_ they require (e.g. android_gingerbread_mr1 and android_honeycomb_mr2).
+level`_ they require (e.g. android_10 and android_honeycomb_mr2).
 Each class or feature is defined in the library project that represents the
 minimum version of Android required for it to work.
 
@@ -28,10 +28,10 @@ development of data driven Android UIs (e.g.
 .. _Android API level: http://developer.android.com/guide/appendix/api-levels.html
 .. _View: http://developer.android.com/reference/android/view/View.html
 
-android_gingerbread_mr1 library project
+android_10 library project
 ---------------------------------------
 
-android_gingerbread_mr1 (API level 10) is the lowest API level supported. It
+android_10 (API level 10) is the lowest API level supported. It
 provides the base `Activity`_ (:javadoc:`org.ros.android.RosActivity`) and
 `Service`_ (:javadoc:`org.ros.android.NodeMainExecutorService`) for executing
 and managing the lifecycle of your :javadoc:`org.ros.node.NodeMain`\s.
